# docs/swagger.yaml
definitions:
...
  controllers.jadwalKuliahInput:
    properties:
      dosen_id:
        type: integer
      hari:
        type: string
      jam_mulai:
        type: string
      jam_selesai:
        type: string
      mahasiswa_id:
        type: integer
    required:
    - dosen_id
    - hari
    - jam_mulai
    - jam_selesai
    - mahasiswa_id
    type: object
  controllers.mahasiswaInput:
    properties:
      nama:
        type: string
    required:
    - nama
    type: object
  controllers.mataKuliahInput:
    properties:
      nama:
        type: string
    required:
    - nama
    type: object
...
  models.Dosen:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      jadwalKuliah:
        items:
          $ref: '#/definitions/models.JadwalKuliah'
        type: array
      mataKuliahID:
        type: integer
      nama:
        type: string
      updatedAt:
        type: string
    type: object
  models.JadwalKuliah:
    properties:
      createdAt:
        type: string
      dosenID:
        type: integer
      hari:
        type: string
      id:
        type: integer
      jam_mulai:
        type: string
      jam_selesai:
        type: string
      mahasiswaID:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Mahasiswa:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      jadwalKuliah:
        items:
          $ref: '#/definitions/models.JadwalKuliah'
        type: array
      nama:
        type: string
      nilai:
        items:
          $ref: '#/definitions/models.Nilai'
        type: array
      updatedAt:
        type: string
    type: object
  models.MataKuliah:
    properties:
      createdAt:
        type: string
      dosen:
        items:
          $ref: '#/definitions/models.Dosen'
        type: array
      id:
        type: integer
      nama:
        type: string
      nilai:
        items:
          $ref: '#/definitions/models.Nilai'
        type: array
      updatedAt:
        type: string
    type: object
  models.Nilai:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      indeks:
        type: string
      mahasiswaID:
        type: integer
      mataKuliahID:
        type: integer
      skor:
        type: integer
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /jadwal-kuliah:
    get:
      description: Get a list of jadwalKuliah.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.JadwalKuliah'
            type: array
      summary: Get all jadwalKuliah.
      tags:
      - jadwal kuliah
    post:
      description: Creating a new jadwalKuliah.
      parameters:
      - description: the body to create a new jadwalKuliah
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.jadwalKuliahInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.JadwalKuliah'
      security:
      - BearerToken: []
      summary: Create New jadwalKuliah.
      tags:
      - jadwal kuliah
  /jadwal-kuliah/{id}:
    delete:
      description: Delete a jadwalKuliah by id.
      parameters:
      - description: jadwalKuliah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one jadwalKuliah.
      tags:
      - jadwal kuliah
    get:
      description: Get a jadwalKuliah by id.
      parameters:
      - description: jadwalKuliah id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JadwalKuliah'
      summary: Get jadwalKuliah.
      tags:
      - jadwal kuliah
    patch:
      description: Update jadwalKuliah by id.
      parameters:
      - description: jadwalKuliah id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an jadwalKuliah
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.jadwalKuliahInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JadwalKuliah'
      security:
      - BearerToken: []
      summary: Update jadwalKuliah.
      tags:
      - jadwal kuliah
...
  /mahasiswa:
    get:
      description: Get a list of mahasiswa.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Mahasiswa'
            type: array
      summary: Get all Mahasiswa.
      tags:
      - mahasiswa
    post:
      description: Creating a new mahasiswa.
      parameters:
      - description: the body to create a new mahasiswa
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.mahasiswaInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Mahasiswa'
      security:
      - BearerToken: []
      summary: Create New mahasiswa.
      tags:
      - mahasiswa
  /mahasiswa/{id}:
    delete:
      description: Delete a mahasiswa by id.
      parameters:
      - description: mahasiswa id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one mahasiswa.
      tags:
      - mahasiswa
    get:
      description: Get a mahasiswa by id.
      parameters:
      - description: mahasiswa id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Mahasiswa'
      summary: Get mahasiswa.
      tags:
      - mahasiswa
    patch:
      description: Update mahasiswa by id.
      parameters:
      - description: mahasiswa id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an mahasiswa
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.mahasiswaInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Mahasiswa'
      security:
      - BearerToken: []
      summary: Update mahasiswa.
      tags:
      - mahasiswa
  /mata-kuliah:
    get:
      description: Get a list of mataKuliah.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MataKuliah'
            type: array
      summary: Get all mataKuliah.
      tags:
      - mata kuliah
    post:
      description: Creating a new mataKuliah.
      parameters:
      - description: the body to create a new mataKuliah
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.mataKuliahInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MataKuliah'
      security:
      - BearerToken: []
      summary: Create New mataKuliah.
      tags:
      - mata kuliah
  /mata-kuliah/{id}:
    delete:
      description: Delete a mataKuliah by id.
      parameters:
      - description: mataKuliah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one mataKuliah.
      tags:
      - mata kuliah
    get:
      description: Get a mataKuliah by id.
      parameters:
      - description: mataKuliah id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MataKuliah'
      summary: Get mataKuliah.
      tags:
      - mata kuliah
    patch:
      description: Update mataKuliah by id.
      parameters:
      - description: mataKuliah id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an mataKuliah
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.mataKuliahInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MataKuliah'
      security:
      - BearerToken: []
      summary: Update mataKuliah.
      tags:
      - mata kuliah
...
swagger: "2.0"